Check for correct parent account type
